+++ src/components/StarRating/styles.js
@@ -1,0 +1,15 @@
+import { StyleSheet } from 'react-native';
+
+import styleConstants from '../../styleConstants';
+
+const styles = StyleSheet.create({
+  container: {
+    flexDirection: 'row',
+  },
+  icon: {
+    fontSize: 48,
+    color: styleConstants.colors.accent,
+  },
+});
+
+export default styles;